+++ templates/base.html
@@ -6,8 +6,8 @@
   </head>
 
   <body>   
-    <a href={% url "tweets:home" %}>ホーム</a>
-    <a href={% url "accounts:signup" %}>サインアップ</a>
+    <a href="{% url "tweets:home" %}">ホーム</a>
+    <a href="{% url "accounts:signup" %}">サインアップ</a>
     {% block content %}
     {% endblock %}
   </body>
